Rename alice::negotiate to steps
This module includes helper function for each state.
